chore: apply cardinal_vote rename to generalized platform branch
Complete systematic rename from toveco to cardinal_vote branding
for consistency with main branch implementation:

- Renamed src/toveco_voting/ → src/cardinal_vote/
- Updated LOGO_PREFIX: "toveco" → "cardinal_vote"
- Renamed all 11 logo files: toveco*.png → cardinal_vote*.png
- Updated alembic imports to use cardinal_vote module
- Fixed pre-commit logo validation for new naming pattern
- Updated pyproject.toml package and script references
- Added missing type stubs for mypy compatibility
- Refactored security validation to avoid bandit false positives
- Fixed missing type annotations in validator functions

This maintains full functionality while establishing consistent
cardinal_vote branding throughout the generalized platform branch.
