+++ pyproject.toml
@@ -48,7 +48,7 @@ Repository = "https://github.com/toveco/voting.git"
 Issues = "https://github.com/toveco/voting/issues"
 
 [project.scripts]
-toveco-voting = "toveco_voting.main:main"
+cardinal-vote = "cardinal_vote.main:main"
 
 [project.optional-dependencies]
 dev = [
@@ -72,7 +72,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/toveco_voting"]
+packages = ["src/cardinal_vote"]
 
 [tool.ruff]
 target-version = "py311"
