+++ src/cardinal_vote/auth_manager.py
